+++ packages/analytics-js/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.28.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.27.0...@rudderstack/analytics-js@3.28.0) (2026-01-22)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-cookies` updated to version `0.5.6`
+* `@rudderstack/analytics-js-common` updated to version `3.27.0`
+* `@rudderstack/analytics-js-plugins` updated to version `3.15.0`
+
+### Features
+
+* reduce noise in error reporting ([#2696](https://github.com/rudderlabs/rudder-sdk-js/issues/2696)) ([78d98e7](https://github.com/rudderlabs/rudder-sdk-js/commit/78d98e7ccbc58327ff94ccaa9c9c669f14fa6a11))
+
 ## [3.27.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.26.0...@rudderstack/analytics-js@3.27.0) (2026-01-12)